+++ changelog.md
@@ -4,6 +4,7 @@
 
 #1.0.3
 - ore shrubs integration (https://minecraft.curseforge.com/projects/ore-shrubs)
+- extra utilities 2 integration (ender lilly and red orchid)
 
 #1.0.2
 - moved to MMD servers